+++ controllers/user-controller.js
@@ -1,4 +1,4 @@
-const { User, Tweet, Reply, Like, Followship } = require('../models')
+const { User, Tweet, Reply, Like } = require('../models')
 const { getUser } = require('../_helpers')
 const bcrypt = require('bcryptjs')
 
@@ -78,52 +78,7 @@ const userController = {
   },
   getReply: (req, res) => {
     res.render('modals/reply')
-  },
-  addFollowing: async (req, res, next) => {
-    try {
-      const id = req.params.id || req.body.id
-      const loginUserId = getUser(req) && getUser(req).id
-
-      if (id === loginUserId.toString()) {
-        return res.redirect('back')
-      }
-
-      const user = await User.findByPk(id)
-      if (!user) throw new Error("User didn't exist!")
-
-      const followship = await Followship.findOne({
-        where: {
-          followerId: loginUserId,
-          followingId: id
-        }
-      })
-      if (followship) throw new Error('You are already following this user!')
-
-      await Followship.create({
-        followerId: loginUserId,
-        followingId: id
-      })
-      return res.redirect('back')
-    } catch (err) {
-      next(err)
-    }
-  },
-  removeFollowing: async (req, res, next) => {
-    try {
-      const followship = await Followship.findOne({
-        where: {
-          followerId: getUser(req) && getUser(req).id,
-          followingId: req.params.id
-        }
-      })
-      if (!followship) throw new Error("You haven't followed this user!")
-      await followship.destroy()
-      return res.redirect('back')
-    } catch (err) {
-      next(err)
-    }
   }
 }
 
 module.exports = userController
-
