+++ views/partials/right.hbs
@@ -8,12 +8,12 @@
     {{!-- {{#each topUsers}} --}}
     <div class="user-info-item d-flex">
       <div class="d-flex align-items-center">
-        <a href="/users/{{this.id}}/tweets">
+        <a href="/users/tweets/{{this.id}}">
           <img src="/images/alphacamp.png" alt="" class="user-avatar m-3" style="width:50px;height:50px">
         </a>
       </div>
       <div class="align-self-center">
-        <a href="/users/{{this.id}}/tweets" style="text-decoration: none;">
+        <a href="/users/tweets/{{this.id}}" style="text-decoration: none;">
           <p class="user-name m-0" style="color: #171725;font-size: 16px;">Apple</p>
         </a>
         <p class="user-name m-0" style="color: #92929D;font-size: 14px;">@apple</p>
